+++ src/lambda/commands/list_for_user.py
@@ -4,14 +4,41 @@ import os
 
 def list_command(user_id):
     query = """
-        select user_commands.id
-        ,commands.text
-        ,user_commands.disabled
-        ,commands.id
-        from commands
-        left join user_commands on commands.id = user_commands.command_id
-        where user_commands.user_id = {user_id}
-        or user_commands.user_id is null
+        select uc.id
+                ,uc.text
+                ,uc.enabled
+                ,uc.id
+        from (
+                select u.id as id
+                ,c.text as text
+                ,u.enabled as enabled
+                ,c.id as command_id
+                from commands c
+                left join user_commands u on c.id = u.command_id
+                where c.id not in (
+                        select u.command_id
+                        from user_commands u
+                        where u.user_id = {user_id}
+                )
+                union
+                select u.id as id
+                ,c.text as text
+                ,u.enabled as enabled
+                ,c.id as command_id
+                from commands c
+                left join user_commands u on c.id = u.command_id
+                where c.id in (
+                        select u.command_id
+                        from user_commands u
+                        where u.user_id = {user_id}
+                )
+                and u.user_id = {user_id}
+        ) as uc
+        order by (
+                case when enabled then 1
+                when enabled is null then 2
+                else 3 end
+        ) asc, text asc
     """.format(user_id=user_id)
     print(query)
     if os.getenv('DEV'):