+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Changed
 
 * List of items on/off individualized per user
+* New commands will show up for users (disabled) immediately.
 
 # 4.0.8
 ## Changed